third_party/WebKit/LayoutTests/fast/files/url-null.html

<!DOCTYPE HTML>
<html>
 <head>
    <script src="../../resources/js-test.js"></script>
 </head>
 <body>
    <div id="description"></div>
    <div id="console"></div>
<script>
description("Test URL methods with null and undefined arguments.");

shouldThrow("URL.createObjectURL(null)");
shouldThrow("URL.createObjectURL(undefined)");
shouldNotThrow("URL.revokeObjectURL(null)");
shouldNotThrow("URL.revokeObjectURL(undefined)");
var successfullyParsed = true;

</script>
 </body>
</html>
